Split tests in eager and non-eager mode
